+++ Principes/Principes/Open_Closed/Correct/Shape.cs
@@ -2,6 +2,6 @@
 {
     public abstract class Shape
     {
-        public abstract int Area();
+        public abstract double Area();
     }
 }
